requirements.txt
selenium~=4.1.0
python-telegram-bot
requests==2.27.1